Move tests into single file
- Remove 'RandomX' classes, prefer methods
